Fix environment check basic auth tests
Was assuming that Member::currentMember() is set,
which isn't the case unless you log in beforehand and start a PHP session.

Given we want the checks to run on their own without user interaction,
they need to work in isolation through a single HTTP call containing
the basic auth info.